DRAFT 강의는 개설자만 조회 가능하도록 접근 제어 추가, price를 원 단위 정수로 재모델링

+++ src/main/java/com/liveklass/enrollment/course/dto/CourseResponse.java
@@ -3,14 +3,13 @@ package com.liveklass.enrollment.course.dto;
 import com.liveklass.enrollment.course.Course;
 import com.liveklass.enrollment.course.CourseStatus;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CourseResponse(
         Long id,
         String creatorId,
         String title,
-        BigDecimal price,
+        Long price,
         int capacity,
         LocalDate startDate,
         LocalDate endDate,
